fix(V2): BF16-to-NumPy .float() conversion + overseer report Run 4 progress
- Added .float() before .cpu().numpy() in 3 files to fix BFloat16
  tensor incompatibility with NumPy:
  - train_v2_backbone.py: _collect_val_predictions() line 681
  - oof_generator.py: line 267
  - stage1_individual_v2.py: lines 138, 149
- Updated PROJECT_OVERSEER_REPORT_DISEASE.md with Run 4 results:
  - 14/15 backbones complete, zero NaN, zero BF16 warnings
  - Top: DynamicConvNet 95.85%, CSPDarkNet 95.76%
  - Swin improved 87->91.23%, ViTHybrid Phase B at ~91%
  - MaxViT 77.10%, CoAtNet 78.32% (architectural ceiling)
  - Health: Yellow->Green

+++ V2_segmentation/ensemble_v2/stage1_individual_v2.py
@@ -135,7 +135,7 @@ class Stage1IndividualV2:
                 cls_logits = outputs.get("cls_logits")
                 if cls_logits is not None:
                     cls_probs = F.softmax(cls_logits, dim=1)
-                    all_cls_probs.append(cls_probs.cpu().numpy())
+                    all_cls_probs.append(cls_probs.float().cpu().numpy())
 
                 # Segmentation masks
                 seg_logits = outputs.get("seg_logits")
@@ -146,7 +146,7 @@ class Stage1IndividualV2:
                         seg_probs, size=(56, 56),
                         mode="bilinear", align_corners=False,
                     )
-                    all_seg_masks.append(seg_probs.cpu().numpy())
+                    all_seg_masks.append(seg_probs.float().cpu().numpy())
 
                 all_labels.append(labels.numpy())
